Pragmatic parsing of initial RCS revision
Usually, RCS will ask for a log message when checking in a new revision. Except
when checking in the first revision, in which case it will ask for a
description - clearly stating that this is NOT the log message, but instead of
asking for a log message afterwards, just putting in the fixed string "Initial
revision". Thus, the description will commonly contain information relevant to
the initial revision, while its log message contains no useful information
whatsoever.

Since the fast-import format does not have an equivalent of RCS's description
string, we take the pragmatic approach of patching the parsing code to check
for "Initial revision" log messages, and replace them with the file's
description string.