main: Grab memory in bigger chunks
The default sbrk call grabs memory from the OS in 128kb chunks for malloc.
As we use up memory in prodigous amounts it is better to grab it in
bigger chunks, and make less calls to sbrk. 16MB seems to be a decent value.

Whether or not this will work with transparent hugepages is a good
question.