Add --embed-id option to edit commit message.
This adds the CVS revision numbers of the changes in this commit as
extra lines on the end of the commit message. Each revision identifier
is preceded by a 'CSV-ID:' tag to allow easy searching.

Putting the old CVS revision identifiers in the git repository in this
way is preferable to attempting to find and replace all the old references
because that is simply impossible. It may be that all the references
within the repository can be found, but if the application has ever been
published it becomes impossible to guess where references might exist,
let alone fix them.

In addition it provides an unambiguous identification as to which commits
originated in CVS in the event that they are 'cherry-picked', rebased or
otherwise duplicated.