Add --expand=k command line option.
If the "-k" option or "--expand" without an argument is given the
tokens will be expanded in the standard form: $key: value$

The default is to copy the $Id$ tags unchanged from the CVS file as
this is faster than searching for the keywords.

This is normally the same as leaving them as "$Id$", BUT sometimes
the repo has these keywords expanded INSIDE THE REPOSITORY.
For these cases the option "--expand=k" will trim the tags down to
the unexpanded form even if they are expanded in the repository.

All the forms from the 'expand_override' function are accepted; though
probably not useful.

--expand=kv	default form, $<key>: <value>$
--expand=kvl	like KKV but with locker's name inserted
--expand=k	keyword-only expansion, $<key>$
--expand=v	value-only expansion, $<value>$
--expand=o	old-value expansion
--expand=b	old-value with no EOL normalization