Check all fixed points against the CVS CVS repo.
This script connects to The CVS repo on savannah.nongnu.org using rsync to
collect the repository. It then converts this to git using the executable
from the above directory and git-fast-import.

Once the repo is converted every reasonable branch and tag is checkout
out of both repositories in sequence and compared.  Any significant
differences are recorded.