Use interned cvs_number pointers
It seems more natural once we've interned the cvs_number pointers
to use them as keys in the hash table as well as passing things
as pointers instead of by value.

Also separate out two functions that play with the version keys
I'm not sure they are used any more.