Speed up merge_branches
Maintain a sort buffer of the revisions in date order
for ease of finding cliques. After an initial sort, try
and minimize future sort ops.

The easy way of just sorting the revisions array works
as far as speeding up merge goes, but it kills the fast
out performance (it breaks the assumptions of PRESORT)

the naive sorting method cut the netbsd-pkgsrc merge from
~3300s to ~1700s. With the additional machinery to avoid
killing fast out, it is more modest. down to ~2900s

On emacs historic, merge went down from 21s to 16s