Cache results of strncmp between directory names
Create master_dir as a separate struct use it to store name,
strlen(name) and a hash value of directories.

Add a master_dir reference to rev_master

Populate the master_dir, atomically so masters with the same
directory have the same reference.

Cache the results of checking whether one directory is an ancestor
of another.

NetBSD calls this check about 7,000,000,000 times, but with only
95,000 different parameters.