Optimize back linking commits.
Change git_commit->dirs from being embedded in the struct to being
a pointer.

This means we know the size of a commit before doing building the
revdirs, which in turn means we can allocate the back links at the
same time we are building the file array.

We lose some as we now have an extra dereference to get the array
members. However, on netbsd-src it saved about 3.5 minutes out of
31 minutes.

The two inner loops around revisions was 26% of the total CPU time
before this change.

This also opens the opportunity to slab allocate git commits.
Probably backwards would be advantageous.

The additional pointers aren't freed. The place to do so would be
just before we tear the process down.