Make iterators the interface to revdir usage.
It is intended that revdir be an opaque type, however we lose too much
performance if we can't take advantage of struct packing in cvs.h

However, you should only build one through revdir_pack_files and only
access one using the revdir_iter functions.

I can imagine an additional interface to create a revdir in a streaming
manner would also be useful in the future.