Support tag subsets if SUBSETTAG defined
In CVS you can tag a subset of files, as well as a whole repo.

Once we've found the join point between a tag and the git repo,
check whether the file states are the same.

If so, then do as now, create a git tag pointing to the commit.

If not, create a branch instead of a tag, parented at the join
point with the exact file contents of the tag.

For large repos, you wil need to increase the size of
branchcount_t.