Major simplification and improvement of vendor branch handling.
The new logic passes all regression tests and correctly handles edge cases the
old did not, as demonstrated by the oldhead repo.

However, it does require -F on the t9601 test's conversion options,
othewise that test triggers a fatal error due to out-of-order-commit.
Not trying to fix this yet because that test is extremely pathological.