Test Case for Problem in new Vendor Logic
The tag "lost-repo-state" should check out files with the following
content:

FILE1 2
FILE2 2

Note, the problem is not that the tag gets assigned to the wrong
commit, but that there is no right commit. This repo state does
not exist in the converted repo.

This used to get the correct result