Fix Tag Bug
We have a fast heueristic that finds the most likely git commit G
to point a tag T at. The tagging logic is now

1. Use the heuristic to find G
2. Check whether the revisions in G are identical to T
3. If so, we're done
4. if not, check all git commits for equivalent revisions (pruning if
   possible)
5. If we find one we're done

What happens next depends on SUBSETTAG

If enabled, we create a branch at G with a single commit containing
the correct set of revisions.

If not we just use G as the tag point as per the previous code.

This is _much_ slower than just using the heuristic method.

There seem to be issues with old cvs using 1.1 and 1.1.1.1
interchangeably